Create README.md
## MANAGER DATABASE
This is a project in test of development that has not yet finished, but it is already quite useful, because it allows to review the database of any type that supports the codeIgniter framework. This small manager is agile and extensible. This project must be finished yet. Please, if you are interested in contributing, make a fork to download the code.